+++ src/py42/sdk/__init__.py
@@ -57,19 +57,6 @@ class SDKClient(object):
         return cls(sdk_dependencies)
 
     @property
-    def storageaccess(self):
-        """A collection of methods for creating StorageClients, which are useful for interacting
-        with storage nodes directly. For most cases, you can be successful using `py42.sdk.archive`
-        or `py42.sdk.securitydata` which abstract away much of the necessary work you have to do
-        when creating and using your own StorageClients. However, `storageaccess` is useful for
-        when you need to be more granular when working with storage nodes.
-
-        Returns:
-            :class:`py42.clients.storage.StorageClientFactory`
-        """
-        return self._sdk_dependencies.storage_client_factory
-
-    @property
     def serveradmin(self):
         """A collection of methods for getting server information for on-premise environments
         and tenant information for cloud environments.