+++ tests/_internal/test_client_factories.py
@@ -1,23 +1,14 @@
 import pytest
 
-from py42._internal.token_providers import StorageTokenProviderFactory
-from py42.clients import (
-    administration,
-    archive,
-    devices,
-    file_event,
-    legalhold,
-    orgs,
-    securitydata,
-    storage,
-)
-from py42.clients import alerts, detectionlists, users
-from py42._internal import key_value_store
 from py42._internal.client_factories import AuthorityClientFactory, MicroserviceClientFactory
-from py42.clients.storage import StorageClientFactory
-
+from py42._internal.clients import key_value_store, storage
+from py42._internal.clients.storage import StorageClientFactory
 from py42._internal.session_factory import SessionFactory
 from py42._internal.storage_session_manager import StorageSessionManager
+from py42._internal.token_providers import StorageTokenProviderFactory
+from py42.clients import administration, archive, devices, file_event, legalhold, orgs
+from py42._internal.clients import securitydata
+from py42.clients import alerts, detectionlists, users
 from py42.sdk.usercontext import UserContext
 
 _USER_UID = "user-uid"