+++ CHANGELOG.md
@@ -16,11 +16,14 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
     instead of ("ACTIVE", "INACTIVE", or "ALL") on the following `LegalHoldClient` methods:
     - `get_all_matters()`
     - `get_all_matter_custodians()`
+- Parameter `storageaccess` was removed from `SDKClient`. To restore files, just use
+    `SDKClient.archive.stream_from_backup()`.
 
 ### Removed
 
-- Parameter `classification` removed from `OrgClient.create_org()`
-- Parameter `legal_hold_membership_uid` removed from `LegalHoldClient.get_all_matter_custodians()`
+- Parameter `classification` removed from `OrgClient.create_org()`.
+- Parameter `legal_hold_membership_uid` removed from `LegalHoldClient.get_all_matter_custodians()`.
+- Removed `SecurityClient`. Use `SecurityModule`.
 
 ### Added